fix(desktop): stop new triggers from silently matching nothing
From CodeRabbit's review of this PR.

newGithubConfig left branches and labels null whenever the event showed
those slots. Null matches nothing — deliberately, so a half-built trigger
cannot fire on every repository — but the chip renders null with its empty
label, which for these reads "Any branch" and "Any label". Nothing
validates them either, so a new "New push to branch" trigger looked
complete, claimed to match every branch, and would never have fired.

They now start at "any", shown or not: they are optional narrowings, unlike
the repository, which stays null and is what the form refuses to save
without. Clearing one back to nothing means "any" for the same reason —
the chip says so either way, and null would make it a lie.

Also from that review: "PR pushed to in" was not a sentence, and each
schedule's next run was formatted in the automation-level timezone, which
is the soonest schedule's — wrong for every other row once schedules can
sit in different zones. The timezone now travels with the date.

chips.tsx held five components against the one-per-file rule in AGENTS.md;
they are now co-located components with their shared classes in
chipStyles.ts.

Verified against the dev database: a saved push trigger stores
branches {"mode":"any"} rather than null.
