+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/automations/components/TriggerSentence/components/ScopeChip/index.ts
@@ -1,0 +1,1 @@
+export { ScopeChip } from "./ScopeChip";
