+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/automations/components/TriggerSentence/components/SelectChip/index.ts
@@ -1,0 +1,1 @@
+export { SelectChip } from "./SelectChip";
