+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/automations/components/TriggerSentence/components/TextFilterChip/index.ts
@@ -1,0 +1,1 @@
+export { TextFilterChip } from "./TextFilterChip";
