+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/automations/components/TriggerSentence/index.ts
@@ -1,3 +1,3 @@
-export type { ScopeOption } from "./chips";
+export type { ScopeOption } from "./scopeOption";
 export { GITHUB_MENU, newGithubConfig } from "./sentence";
 export { TriggerSentence } from "./TriggerSentence";
