+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/automations/components/TriggerSentence/scopeOption.ts
@@ -1,0 +1,2 @@
+/** One selectable value in a scope or actor chip. */
+export type ScopeOption = { id: string; label: string };
